+++ internal/clients/connect/client_connect.go
@@ -562,3 +562,13 @@ func (c *ConnectClient) GetCurrentUser(log logging.Logger) (*User, error) {
 	// Convert from UserDTO to User
 	return connectUser.toUser(), nil
 }
+
+func (c *ConnectClient) GetIntegrations(log logging.Logger) ([]Integration, error) {
+	var integrations []Integration
+	err := c.client.Get("/__api__/v1/oauth/integrations", &integrations, log)
+	if err != nil {
+		return nil, err
+	}
+	return integrations, nil
+}
+